fix(automation): reject editing items with invalid JSON fields
Fail closed in dtoToEditor on bad schedule/conditions/actions JSON so
corrupt data cannot be silently rewritten, and serialize run-log loads.
