+++ frontend/src/i18n/locales/ko.json
@@ -318,6 +318,7 @@
     "saveAndContinue": "저장 후 계속",
     "discardAndContinue": "버리고 계속",
     "saveError": "저장에 실패했습니다",
+    "itemParseError": "데이터가 손상되어 안전하게 편집할 수 없습니다",
     "toggleError": "전환에 실패했습니다",
     "deleteConfirm": "「{name}」을(를) 삭제할까요?",
     "weekday0": "일",
