+++ frontend/src/i18n/locales/zh-CN.json
@@ -318,6 +318,7 @@
     "saveAndContinue": "保存并继续",
     "discardAndContinue": "放弃并继续",
     "saveError": "保存失败",
+    "itemParseError": "数据损坏，无法安全编辑",
     "toggleError": "切换失败",
     "deleteConfirm": "要删除「{name}」吗？",
     "weekday0": "日",
