+++ frontend/src/i18n/locales/zh-TW.json
@@ -318,6 +318,7 @@
     "saveAndContinue": "儲存並繼續",
     "discardAndContinue": "捨棄並繼續",
     "saveError": "儲存失敗",
+    "itemParseError": "資料損毀，無法安全編輯",
     "toggleError": "切換失敗",
     "deleteConfirm": "要刪除「{name}」嗎？",
     "weekday0": "日",
